Update index of Lambda subnet
1st IP block is used by the public and private subnets. So start with the 2nd IP range.

There is actually logic between the subnet sizes making it really complex to make it "perfect".

This solution can result resul in errors if: Public, private and Lambda subnets all have different sized. But also if the Lambda subnet is really small (e.g. a `/23` when the private/public are `/24`)